# src/main/resources/布灵4.0.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "布灵",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/user/collections": {
      "get": {
        "summary": "获取当前用户的收藏夹分组列表",
        "deprecated": false,
        "description": "",
        "operationId": "listMyCollectionGroups",
        "tags": [
          "收藏夹 (Collections)"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultCollectionGroupList"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "创建新的收藏夹分组",
        "deprecated": false,
        "description": "",
        "operationId": "createCollectionGroup",
        "tags": [
          "收藏夹 (Collections)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionGroupCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultCollectionGroup"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/collections/{groupId}": {
      "delete": {
        "summary": "删除收藏夹分组",
        "deprecated": false,
        "description": "",
        "operationId": "deleteCollectionGroup",
        "tags": [
          "收藏夹 (Collections)"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "要删除的分组ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/collections/{groupId}/articles": {
      "get": {
        "summary": "获取指定分组下的收藏文章列表",
        "deprecated": false,
        "description": "",
        "operationId": "listArticlesInCollectionGroup",
        "tags": [
          "收藏夹 (Collections)"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "收藏夹分组ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "状态码 (200:成功, 其他:失败)",
                      "examples": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息",
                      "examples": [
                        "Success"
                      ]
                    },
                    "data": {
                      "type": "object",
                      "description": "通用的历史记录对象结构",
                      "properties": {
                        "resourceId": {
                          "type": "string",
                          "description": "文档的ID"
                        },
                        "resourceTitle": {
                          "type": "string",
                          "description": "文档的标题"
                        },
                        "resourceType": {
                          "type": "string",
                          "description": "文档的类型 (e.g., markdown)"
                        },
                        "kbId": {
                          "type": "string",
                          "description": "所属知识库的ID"
                        },
                        "kbName": {
                          "type": "string",
                          "description": "所属知识库的名称"
                        },
                        "authorId": {
                          "type": "string",
                          "description": "文档作者的用户ID"
                        },
                        "authorName": {
                          "type": "string",
                          "description": "文档作者的昵称"
                        },
                        "actionAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "操作发生的时间 (UTC)"
                        }
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "将文章添加到收藏夹",
        "deprecated": false,
        "description": "",
        "operationId": "addArticleToCollectionGroup",
        "tags": [
          "收藏夹 (Collections)"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "收藏夹分组ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionItemCreateRequest"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "201": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/collections/{groupId}/articles/{articleId}": {
      "delete": {
        "summary": "从收藏夹移除文章",
        "deprecated": false,
        "description": "",
        "operationId": "removeArticleFromCollectionGroup",
        "tags": [
          "收藏夹 (Collections)"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "收藏夹分组ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "articleId",
            "in": "path",
            "description": "要移除的文章ID (resourceId)",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200:成功, 其他:失败)",
            "examples": [
              200
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "examples": [
              "Success"
            ]
          },
          "data": {
            "type": "object",
            "description": "响应数据",
            "properties": {},
            "nullable": true
          }
        }
      },
...
      "CollectionGroupCreateRequest": {
        "type": "object",
        "required": [
          "groupName"
        ],
        "properties": {
          "groupName": {
            "type": "string",
            "description": "新分组的名称"
          }
        }
      },
      "CollectionItemCreateRequest": {
        "type": "object",
        "required": [
          "resourceId"
        ],
        "properties": {
          "resourceId": {
            "type": "string",
            "description": "要收藏的文章ID"
          }
        }
      },
      "CollectionGroup": {
        "type": "object",
        "description": "收藏夹分组信息",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupName": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "description": "分组内收藏的文章数量"
          }
        }
      },
...
      "ResultCollectionGroup": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CollectionGroup"
              }
            }
          }
        ]
      },
      "ResultCollectionGroupList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CollectionGroup"
                }
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}